+++ WeatherApp/WeatherApp/Controllers/HomeViewController/HourlyDisplayData.swift
@@ -1,0 +1,33 @@
+//
+//  HourlyDisplayData.swift
+//  WeatherApp
+//
+//  Created by Phan Quyen on 18/08/2025.
+//
+
+import Foundation
+
+struct HourlyDisplayData {
+    let time: String
+    let temperature: String
+    let icon: String
+    
+    var weatherIcon: String {
+        switch icon {
+        case "sun.max":
+            return "○"
+        case "cloud.sun":
+            return "◐"
+        case "cloud":
+            return "●"
+        case "cloud.rain":
+            return "¤"
+        case "moon":
+            return "◑"
+        case "snow":
+            return "❋"
+        default:
+            return "○"
+        }
+    }
+}
