+++ WeatherApp/WeatherApp/Controllers/HomeViewController/WeatherDisplayData.swift
@@ -1,5 +1,5 @@
 //
-//  WeatherDisplayModels.swift
+//  WeatherDisplayData.swift
 //  WeatherApp
 //
 //  Created by Phan Quyen on 18/08/2025.
@@ -11,19 +11,6 @@ struct WeatherDisplayData {
     let cityName: String
     let temperature: String
     let description: String
-    let high: String
-    let low: String
-    let icon: String
-}
-
-struct HourlyDisplayData {
-    let time: String
-    let temperature: String
-    let icon: String
-}
-
-struct WeeklyDisplayData {
-    let day: String
     let high: String
     let low: String
     let icon: String
